Fix enable.sh on CoreOS
I spoke briefly with robszumski in the #coreos channel on
Freenode.  He suggested the appropriate place to install Docker Compose
was /opt/bin.  This directory doesn't exist by default on CoreOS so this
change includes creation of the directory. By default this directory
does, however exist in the $PATH variable. Since this is an installation
of an application, I've also moved it to install.sh instead of
enable.sh.  I'm also bumping the Compose version number from 1.1.0-rc2
to 1.2.0