pkg/vmextension: more refactoring for usability
We now provide a ReadSettings(dir) method so that extensions can
reach to JSON objects (as map[string]interface{}) before we unmarshal
them into structs.

This way they can convert the json object back to a string and
run validation (e.g. JSON-schema) on it. Then they can give it
back to UnmarshalHandlerSettings to parse it into actual
structs.