Generate Docker certificates at test time
The integration tests did not work because the test certificates had
expired on 12/05/2016.  Certificates and the extension configuration are
now auto-generated at test time to prevent this from happening ever
again.

The images specified are out of date, and had to be updated.  This is
equally bad, and will be correcte in a future commit.